Refactored namespace for IConverterFactory
